move tests from test/python/transpiler/aqc to test/python/synthesis/aqc
