+++ test/python/synthesis/aqc/test_cnot_networks.py
@@ -12,7 +12,7 @@
 """
 Tests building up CNOT unit structures.
 """
-from test.python.transpiler.aqc.sample_data import CARTAN_4, CARTAN_3
+from test.python.synthesis.aqc.sample_data import CARTAN_4, CARTAN_3
 
 import numpy as np
 from ddt import ddt, data, unpack
